Allow for bigger dummy column price

+++ src/preprocessing/program_params.cpp
@@ -15,7 +15,7 @@ ProgramParams::ProgramParams(const std::string& data_file_name) {
     delta = pt.get<int>("delta");
     theta = pt.get<int>("theta");
     max_cols_to_solve_mp = pt.get<int>("max_cols_to_solve_mp");
-    dummy_column_price = pt.get<int>("dummy_column_price");
+    dummy_column_price = pt.get<long>("dummy_column_price");
     cplex_cores = pt.get<int>("cplex_cores");
     elementary_labelling_every_n_nodes = pt.get<int>("elementary_labelling_every_n_nodes");
     try_elementary_labelling = pt.get<bool>("try_elementary_labelling");
